requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cryptography==3.4.7
DateTime==4.3
Deprecated==1.2.12
distro==1.5.0
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.17.0
fastapi==0.66.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
h11==0.12.0
idna==3.2
joblib==1.0.1
jsonschema==3.2.0
numpy==1.21.1
opentelemetry-api==1.4.0
opentelemetry-exporter-jaeger==1.4.0
opentelemetry-exporter-jaeger-proto-grpc==1.4.0
opentelemetry-exporter-jaeger-thrift==1.4.0
opentelemetry-instrumentation==0.23b0
opentelemetry-instrumentation-asgi==0.23b0
opentelemetry-instrumentation-fastapi==0.23b0
opentelemetry-sdk==1.4.0
opentelemetry-semantic-conventions==0.23b0
opentelemetry-util-http==0.23b0
paramiko==2.7.2
passlib==1.7.4
protobuf==3.17.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
starlette==0.14.2
texttable==1.6.4
threadpoolctl==2.2.0
thrift==0.13.0
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
websocket-client==0.59.0
wrapt==1.12.1
zope.interface==5.4.0
